Create and use test utils to test web components
