+++ pkg/app/web/src/components/environment-list-item.stories.tsx
@@ -3,9 +3,7 @@ import { EnvironmentListItem } from "./environment-list-item";
 
 export default {
   title: "EnvironmentListItem",
-  component: EnvironmentListItem
+  component: EnvironmentListItem,
 };
 
-export const overview: React.FC = () => (
-  <EnvironmentListItem />
-);
+export const overview: React.FC = () => <EnvironmentListItem id="env-id" />;
